+++ src/Skoruba.Duende.IdentityServer.Admin.EntityFramework.MySql/Helpers/MigrationAssembly.cs
@@ -1,7 +1,0 @@
-﻿namespace Skoruba.Duende.IdentityServer.Admin.EntityFramework.MySql.Helpers
-{
-    public class MigrationAssembly
-    {
-        
-    }
-}